converting to 1 page layout
